Separate close button into own file
